+++ CHANGES.md
@@ -1,6 +1,7 @@
 v0.3
 --------------
 - removed authc.authc_account module
+- removed all traces of DefaultCompositeAccount from authc strategies
 - changed remembered-or-authenticated validation logic used during authorization
 - expanded mgt.login process to support multi-factor, creating a new subject only
   after all authentication is complete
@@ -10,6 +11,11 @@ v0.3
     - added support for a "MFA Challenger" (such as SMS, GoogleAuthenticator, etc)
 - authenticating realms now feature a "supported_tokens" attribute that returns
   collection of authentication token types supported
+- removed the Resolver design pattern:  credential, authz_info, permission, role
+- introduced lock-account workflow between Authenticator, AccountStoreRealm, and AccountStore
+- removed setter side effects during realms setting
+- updated cache key domains for authentication and authorization (now realm-specific)
+- broadened accountstore api from get_credentials to get_authc_info
 
 Yosai uses [GitHub's Releases feature](https://github.com/blog/1547-release-your-software) for its changelogs.